requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.8.3
click==8.2.1
coverage==7.10.2
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
